requirements.txt
flask
pandas
gunicorn
flask-cors
requests
json
gspread
google.oauth2.service_account
datetime
python-dotenv